fix builds for arm64
