+++ pkg/ebpf/kernelbugs/kernel_bugs.go
@@ -6,24 +6,10 @@
 package kernelbugs
 
 import (
-	"bytes"
 	_ "embed"
-	"fmt"
-	"log"
-	"os/exec"
-	"syscall"
-	"time"
 
 	ddebpf "github.com/DataDog/datadog-agent/pkg/ebpf"
-	"github.com/DataDog/datadog-agent/pkg/ebpf/bytecode/runtime"
-	manager "github.com/DataDog/ebpf-manager"
 )
-
-//go:embed c/uprobe-trigger.o
-var SimpleUretprobe []byte
-
-//go:embed c/detect-seccomp-bug
-var TriggerProgram []byte
 
 // HasTasksRCUExitLockSymbol returns true if the tasks_rcu_exit_srcu symbol is found in the kernel symbols.
 // The tasks_rcu_exit_srcu lock might cause a deadlock when removing fentry trampolines.
@@ -38,69 +24,4 @@ func HasTasksRCUExitLockSymbol() (bool, error) {
 	// VerifyKernelFuncs returns the missing symbols
 	_, isMissing := missingSymbols[tasksRCUExitLockSymbol]
 	return !isMissing, nil
-}
-
-// HasUretprobeSyscallSeccompBug returns true if the running kernel blocks the uretprobe syscall in seccomp
-// This can cause a probed application running within a seccomp context to segfault.
-// https://lore.kernel.org/lkml/CAHsH6Gs3Eh8DFU0wq58c_LF8A4_+o6z456J7BidmcVY2AqOnHQ@mail.gmail.com/
-func HasUretprobeSyscallSeccompBug() (bool, error) {
-	const uretprobeSyscallSymbol = "__x64_sys_uretprobe"
-	missingSymbols, err := ddebpf.VerifyKernelFuncs(uretprobeSyscallSymbol)
-	if err != nil {
-		return false, err
-	}
-
-	_, isMissing := missingSymbols[uretprobeSyscallSymbol]
-	if isMissing {
-		return false, nil
-	}
-
-	pfile, err := runtime.NewProtectedFile("detect-seccomp-bug", "/tmp", bytes.NewReader(TriggerProgram))
-	if err != nil {
-		return false, err
-	}
-
-	m := manager.Manager{
-		Probes: []*manager.Probe{
-			{
-				ProbeIdentificationPair: manager.ProbeIdentificationPair{
-					EBPFFuncName: "uretprobe__segfault",
-				},
-				BinaryPath:    pfile.Name(),
-				MatchFuncName: "trigger_uretprobe_syscall",
-			},
-		},
-	}
-
-	if err := m.Init(bytes.NewReader(SimpleUretprobe)); err != nil {
-		return false, err
-	}
-	defer func() {
-		if err := m.Stop(manager.CleanAll); err != nil {
-			log.Print(err)
-		}
-	}()
-
-	if err := m.Start(); err != nil {
-		return false, err
-	}
-
-	// wait for uprobe to be attached
-	time.Sleep(3 * time.Second)
-
-	cmd := exec.Command(pfile.Name())
-	if err := cmd.Run(); err != nil {
-		if exiterr, ok := err.(*exec.ExitError); ok {
-			exitcode := exiterr.ExitCode()
-			if exitcode == int(syscall.SIGSEGV) {
-				return true, nil
-			} else {
-				return false, fmt.Errorf("unexpected error code %d when probing for uretprobe seccomp bug: %w", exitcode, err)
-			}
-		} else {
-			return false, fmt.Errorf("failed to probe for uretprobe seccomp bug: %w", err)
-		}
-	}
-
-	return false, nil
 }
